Basic setup ready \(^-^)/
Implemented basic structure we are going to use to access the users.

Adi can now start working on her part of the code.

Jeni must work on saving users in a text file, hash the pass, actually generating the id and implementation for checking if user is logined to access things.